feat: heading component prop name change
Changed prop name of the Heading component from "as" to "element" because "as" is a special styled-components prop and if you extend styling of such component, you cannot use directly "as" and you need to use "forwardedAs" which is confusing. You cannot know this unless you read carefully styled-components docs.